refactor(settings-accounts): flatten account-card folders to single files
Collapse google/google-ads/wpcom account-card index.js files into
flat *.js modules; fix relative imports for connected-badge and
account-card-text-detail accordingly. Reorder useAdminUrl import and
prop order in ConnectYouTubeAccountCard; normalize/add JSDoc across
account-card-actions and account cards.
